Disable macro scoping for bootstrap
Turning macro scoping on causes some code to fail. Fixing this
requires major plumbing... *or*, the addition of an "include_macros!"
form that creates a means to include a batch of macros. In order to
solve bootstrap problems, this commit includes the include_macros!
form but doesn't enforce macro scoping; this should make it possible
for the resulting compiler to compile both the old style and the newer
style.